Added StanfordAnnotatorHelperMethods method for collecting possible entity Strings.
StanfordAnnotatorHelperMethods has a method getCorefTypes that returns the NE type (PERSON, ORGANIZATION, LOCATION) of the given KBP Entity and then all the named entities in the document that are of that type.
These candidate entity strings are parsed by helper methods in CorefHelperMethods.scala to choose the best entity string given the type of the entity and the original string.