Added a resource for storing a map from KB Titles to KB IDs
This map is used when we have found a longer string than the original
KBP entity string and the linker linked to nothing in wikipedia. This
is most likely due to a difference between the databses so if there is an
entry in the KB with the same exact string as the larger string
we found in the document, then they should most likely be linked.